# ui/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreviewStatusResponse": {
        "properties": {
          "camera_name": {
            "title": "Camera Name",
            "type": "string"
          },
          "degraded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degraded Reason"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "idle_shutdown_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idle Shutdown At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "state": {
            "$ref": "#/components/schemas/PreviewState"
          },
          "viewer_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewer Count"
          }
        },
        "required": [
          "camera_name",
          "enabled",
          "state"
        ],
        "title": "PreviewStatusResponse",
        "type": "object"
      },
      "PreviewStopResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/PreviewState"
          }
        },
        "required": [
          "accepted",
          "state"
        ],
        "title": "PreviewStopResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HomeSec API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/preview/cameras/{camera_name}": {
      "delete": {
        "description": "Force-stop preview for a camera.",
        "operationId": "force_stop_preview_api_v1_preview_cameras__camera_name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "camera_name",
            "required": true,
            "schema": {
              "title": "Camera Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewStopResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Force Stop Preview",
        "tags": [
          "preview"
        ]
      },
      "get": {
        "description": "Return preview status for a camera.",
        "operationId": "get_preview_status_api_v1_preview_cameras__camera_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "camera_name",
            "required": true,
            "schema": {
              "title": "Camera Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Preview Status",
        "tags": [
          "preview"
        ]
      },
      "post": {
        "description": "Ensure preview is active for a camera and mint a fresh attach token.",
        "operationId": "ensure_preview_active_api_v1_preview_cameras__camera_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "camera_name",
            "required": true,
            "schema": {
              "title": "Camera Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ensure Preview Active",
        "tags": [
          "preview"
        ]
      }
    },
    "/api/v1/preview/cameras/{camera_name}/playlist.m3u8": {
      "get": {
        "description": "Return the live HLS playlist for a camera preview session.",
        "operationId": "get_preview_playlist_api_v1_preview_cameras__camera_name__playlist_m3u8_get",
        "parameters": [
          {
            "in": "path",
            "name": "camera_name",
            "required": true,
            "schema": {
              "title": "Camera Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Preview Playlist",
        "tags": [
          "preview"
        ]
      }
    },
    "/api/v1/preview/cameras/{camera_name}/sessions": {
      "post": {
        "description": "Attach one authorized viewer to the runtime-owned preview.",
        "operationId": "negotiate_preview_session_api_v1_preview_cameras__camera_name__sessions_post",
        "parameters": [
          {
            "in": "path",
            "name": "camera_name",
            "required": true,
            "schema": {
              "title": "Camera Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewOffer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Negotiate Preview Session",
        "tags": [
          "preview"
        ]
      }
    },
    "/api/v1/preview/cameras/{camera_name}/sessions/{session_id}": {
      "delete": {
        "description": "Detach one viewer without stopping other viewers or the camera runtime.",
        "operationId": "close_preview_session_api_v1_preview_cameras__camera_name__sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "camera_name",
            "required": true,
            "schema": {
              "title": "Camera Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewSessionAction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Preview Session",
        "tags": [
          "preview"
        ]
      },
      "patch": {
        "description": "Renew a viewer only for the remaining authenticated lifetime.",
        "operationId": "renew_preview_session_api_v1_preview_cameras__camera_name__sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "camera_name",
            "required": true,
            "schema": {
              "title": "Camera Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewSessionAction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Renew Preview Session",
        "tags": [
          "preview"
        ]
      }
    },
    "/api/v1/preview/cameras/{camera_name}/{segment_name}": {
      "get": {
        "description": "Return a live HLS transport-stream segment for a camera preview session.",
        "operationId": "get_preview_segment_api_v1_preview_cameras__camera_name___segment_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "camera_name",
            "required": true,
            "schema": {
              "title": "Camera Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "segment_name",
            "required": true,
            "schema": {
              "title": "Segment Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Preview Segment",
        "tags": [
          "preview"
        ]
      }
    },
...
  }
}